+++ Makefile
@@ -7,6 +7,7 @@ SRC			=	src/main.cpp \
 				src/createMsgs.cpp \
 				src/connectionMessages.cpp \
 				src/kick.cpp \
+				src/invite \
 				src/part.cpp \
 				src/topic.cpp \
 				src/privmsg.cpp \
